+++ benchmarks/PostQuantum.LMS.Signer.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+using BenchmarkDotNet.Running;
+using PostQuantum.Lms.Benchmarks;
+
+BenchmarkRunner.Run<LmsBenchmarks>();